server/requirements.txt
asgiref==3.5.0
astroid==2.4.2
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
django-jsonfield==1.4.1
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
isort==5.6.4
joblib==1.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
mysqlclient==2.1.0
protobuf==3.19.4
pycodestyle==2.7.0
pylint==2.6.0
pynvim==0.4.3
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
wrapt==1.12.1
